requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sendgrid
openai
